added int face parameter to Texture::update(); It defaults to -1, and when left at default, behavior is unchanged. If set to a value from 0-5, treats face as an index into Texture's static cubeFaceTarget GLenum array. update then copies into the specified cubemap face.
added int face parameter to Texture::getTexImage();
It defaults to -1, and when left at default, behavior is unchanged.
If set to a value from 0-5, treats face as an index into Texture's
static cubeFaceTarget GLenum array. getTexImage then treats the specified
face as the target texture.

add int face parameter to Texture::getImage();
It defaults to -1, and when left at default, behavior is unchanged.
If set to a value from 0-5, treats face as an index into Texture's
static cubeFaceTarget GLenum array. getImage then treats the specified
face as the target texture.

git-svn-id: https://g3d.svn.sourceforge.net/svnroot/g3d@1167 ef0428c2-0c16-0410-9f3f-d710ca1f8061