Added m_cubemapShader and g_cachedCubemapShader to GuiTextureBox. These directly mirror m_shader and g_cachedShader, except are loaded when the texture underlying the GuiTextureBox is a cubemap.
Modified GuiTextureBox::setSettings() to load m_cubemapShader instead
of m_shader into memory when the visualization parameter has
isCubemap set to true. This shader is identical to m_shader, except
treats texture as a samplerCube, and treats the texture coords as
spherical coordinate, using them to calculate a direction vector in
cartesian coordinates to access the samplerCube, instead of indexing
texture directly.

Created GuiTextureBox::setUpShader(Shader::Ref shader) helper function
that sets all the shader args for the specified shader. Now
GuiTextureBox::drawTexture calls this function instead of setting up
the args itself.

Modified GuiTextureBox::drawTexture to use m_shader only on non-cubemaps
and instead use m_cubemapShader on cubemap textures, as specified
by m_visualization.isCubemap. The cubemap is displayed by treating u, v
coordinates as spherical coordinates phi and theta, by multiplying them by
two pi and pi respectively. These spherical coordinates are then translated to
cartesian coodinates

Added m_xyzLabel to GuiTextureBoxInspector, which is assigned only when the
m_settings of the original GuiTextureBox has a true isCubemap
value. Conversely, m_uvLabel is left unassigned in such a case.
This assignment occurs in the constructor.

Modified render to display m_xyzLabel if m_settings.isCubemap is true
and m_uvLabel otherwise. In the cubemap case, the u and v coordinates
are used to determine the xyz access vector into the cubemap, which is
then displayed in the m_xyzLabel

Modified render to double the width of the displayed texture if the texture is a cubemap.

git-svn-id: https://g3d.svn.sourceforge.net/svnroot/g3d@1171 ef0428c2-0c16-0410-9f3f-d710ca1f8061